requirements.txt
shap==0.40.0
joblib==1.1.0
lifelines==0.27.0
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.4
Pillow==9.0.1
seaborn==0.11.2
streamlit==1.7.0
